refactor: rename facetFilters file to match the name of the only exported function in it
